refactor(api): cleanup route definitions
