+++ assets/express/scripts/generate-openapi.ts
@@ -2,33 +2,16 @@
 import '@extensions/zod/register';
 import '@extensions/knex/register';
 
-import pino from 'pino';
 import { writeFileSync } from 'fs';
 import { resolve } from 'path';
 
 import { envSchema } from '@common/environment';
 import { registry } from '@modules';
-import { initializeKnex } from '../src/database/initilize-knex';
-import createDbRepos from '@database';
-import apiDefinitionFactory from '@api';
+import { routes } from '@api';
 import { generateDocument } from '@common/openapi';
 
 const run = async () => {
   const env = Object.freeze(envSchema.parse(process.env));
-
-  const logger = pino({
-    name: 'http',
-    ...(env.NODE_ENV !== 'production' && {
-      transport: {
-        target: 'pino-pretty',
-        colorize: false,
-      },
-    }),
-  });
-
-  const knex = initializeKnex(logger);
-  const dbRepositories = createDbRepos(knex);
-  const routes = apiDefinitionFactory(dbRepositories);
 
   const document = generateDocument(registry, '3.0.3', routes);
   document.servers = [{ url: `http://localhost:${env.PORT}` }];
