+++ assets/express/src/@types/express/index.d.ts
@@ -1,0 +1,9 @@
+import { Services } from '@app/modules';
+
+declare global {
+  namespace Express {
+    interface Request {
+      services: Services;
+    }
+  }
+}
