+++ assets/express/src/api/middleware/index.ts
@@ -1,30 +1,58 @@
-import { json, RequestHandler, ErrorRequestHandler } from 'express';
-import helmet from 'helmet';
-import cors from 'cors';
-import compression from 'compression';
+import { RequestHandler, ErrorRequestHandler } from 'express';
+import { expressjwt as jwt } from 'express-jwt';
 import { Logger } from 'pino';
 import createError from 'http-errors';
+import { z } from 'zod';
 import { serviceToHttpErrorMap } from '@common';
-import { validateAccessToken } from './validateAccessToken';
 import { TokensService } from '@modules';
+import { RequestSchema, RequestKey } from '../interfaces';
+import { Services } from '@app/modules';
 
-export * from './validateRequest';
+export const attachServices =
+  (services: Services): RequestHandler =>
+  (req, _res, next) => {
+    req.services = services;
+    next();
+  };
 
-const logRequest = function (logger: Logger): RequestHandler {
-  return function ({ method, path }, _res, next) {
-    // DO NOT use console: https://expressjs.com/en/advanced/best-practice-performance.html#do-logging-correctly
-    logger.info({ method, path });
+export const validateRequest = function (
+  schema: RequestSchema
+): RequestHandler {
+  return function (req, _res, next) {
+    const result = z.object(schema).safeParse(req);
+
+    if (!result.success) {
+      const error = createError.UnprocessableEntity('Bad Data');
+      error.errors = result.error.issues;
+      return next(error);
+    }
+
+    for (const [k, v] of Object.entries(result.data)) {
+      req[k as RequestKey] = v;
+    }
+
     next();
   };
 };
 
-const handleServiceError: ErrorRequestHandler = (err, _req, _res, next) => {
-  const klass = serviceToHttpErrorMap[err.name];
-  const error = klass ? new klass(err.message) : err;
-  next(error);
+/**
+ * Logs the incoming request's method and URL, e.g. `GET /todos/1`
+ */
+export const logRequest = function (logger: Logger): RequestHandler {
+  return function ({ method, url }, _res, next) {
+    logger.info(`${method} ${url}`);
+    next();
+  };
 };
 
-const errorHandler = function (logger: Logger): ErrorRequestHandler {
+export const handleServiceError =
+  (): ErrorRequestHandler => (err, _req, _res, next) => {
+    const klass = serviceToHttpErrorMap[err.name];
+    const error = klass ? new klass(err.message) : err;
+    next(error);
+  };
+
+export const errorHandler = function (logger: Logger): ErrorRequestHandler {
   return function (err, _req, res, next) {
     // https://expressjs.com/en/guide/error-handling.html
     // If you call next() with an error after you have started writing the response (for example, if you encounter an error while streaming
@@ -48,18 +76,12 @@ const errorHandler = function (logger: Logger): ErrorRequestHandler {
   };
 };
 
-export const initializeMiddlewares = function (
-  logger: Logger,
-  tokensService: TokensService
-) {
-  return [
-    logRequest(logger),
-    helmet(),
-    cors(),
-    json(),
-    compression(),
-    handleServiceError,
-    errorHandler(logger),
-    validateAccessToken(tokensService),
-  ];
+const exceptionPaths: string[] = ['/', '/v1/auth/login', '/v1/auth/register'];
+
+export const validateAccessToken = function (tokensService: TokensService) {
+  const jwtConfig = tokensService.getJwtConfig();
+
+  return jwt({ secret: jwtConfig.secret, algorithms: ['HS256'] }).unless({
+    path: exceptionPaths,
+  });
 };
