+++ assets/express/src/api/v1/auth/routes/index.ts
@@ -1,14 +1,6 @@
-import { AuthService } from '@modules';
 import { prefixRoutes } from '../../../utils';
 
-import loginRouteOptions from './login';
-import registerRouteOptions from './register';
+import login from './login';
+import register from './register';
 
-export default function ({ authService }: { authService: AuthService }) {
-  const routes = [
-    loginRouteOptions({ authService }),
-    registerRouteOptions({ authService }),
-  ];
-
-  return prefixRoutes('/auth', routes);
-}
+export default prefixRoutes('/auth', [login, register]);
