+++ assets/express/src/api/v1/auth/routes/register.ts
@@ -1,27 +1,26 @@
+import { asyncHandler, defineRoute } from '@app/api/utils';
 import { response } from '@common';
-import { models, AuthService } from '@modules';
-import { bindRouteOptionsWithSchema } from '../../../interfaces';
+import { models } from '@modules';
 import { registerDTO } from '../dto';
 
-export default bindRouteOptionsWithSchema(
-  ({ authService }: { authService: AuthService }) => ({
-    operationId: 'register',
-    summary: 'Register a user',
-    description: 'Register a user',
-    tags: ['Auth'],
-    method: 'post',
-    path: '/register',
-    request: {
-      body: registerDTO,
-    },
-    responses: {
-      200: models.Auth,
-      409: response.Conflict(),
-      422: response.UnprocessableEntity(),
-    },
-    handler: async (req, res) => {
-      const tokens = await authService.register(req.body);
-      res.status(200).send(tokens);
-    },
+export default defineRoute({
+  operationId: 'register',
+  summary: 'Register a user',
+  description: 'Register a user',
+  tags: ['Auth'],
+  method: 'post',
+  path: '/register',
+  request: {
+    body: registerDTO,
+  },
+  responses: {
+    200: models.Auth,
+    409: response.Conflict(),
+    422: response.UnprocessableEntity(),
+  },
+}).attachHandler(
+  asyncHandler(async ({ body, services }, res) => {
+    const tokens = await services.authService.register(body);
+    res.status(200).send(tokens);
   })
 );
