+++ assets/express/src/api/v1/index.ts
@@ -1,23 +1,5 @@
-import { DbCollection } from '@database';
-import { defineUserRoutes } from './users';
-import { defineTodoRoutes } from './todos';
-import { defineAuthRoutes } from './auth';
-import {
-  initializeUserService,
-  initializeTodoService,
-  AuthService,
-} from '@modules';
+import users from './users';
+import todos from './todos';
 import { prefixRoutes } from '../utils';
 
-export default function (dbCollection: DbCollection, authService: AuthService) {
-  const userService = initializeUserService(dbCollection);
-  const todoService = initializeTodoService(dbCollection);
-
-  const routes = [
-    ...defineUserRoutes({ userService }),
-    ...defineTodoRoutes({ todoService }),
-    ...defineAuthRoutes({ authService }),
-  ];
-
-  return prefixRoutes('/v1', routes);
-}
+export default prefixRoutes('/v1', [...todos, ...users]);
