--- assets/express/src/api/v1/todos/routes/index.ts
@@ -1,22 +1,0 @@
-import { prefixRoutes } from '../../../utils';
-import { TodoService } from '@modules';
-
-import getTodoRouteOptions from './get-todo';
-import listTodosRouteOptions from './list-todos';
-import createTodoRouteOptions from './create-todo';
-import updateTodoRouteOptions from './update-todo';
-import patchTodoRouteOptions from './patch-todo';
-import deleteTodoRouteOptions from './delete-todo';
-
-export default function ({ todoService }: { todoService: TodoService }) {
-  const routes = [
-    getTodoRouteOptions({ todoService }),
-    listTodosRouteOptions({ todoService }),
-    createTodoRouteOptions({ todoService }),
-    updateTodoRouteOptions({ todoService }),
-    patchTodoRouteOptions({ todoService }),
-    deleteTodoRouteOptions({ todoService }),
-  ];
-
-  return prefixRoutes('/todos', routes);
-}
