+++ assets/express/src/api/v1/users/index.ts
@@ -1,5 +1,17 @@
-import defineUserRoutes from './routes';
+import { prefixRoutes } from '../../utils';
 
-// export * from './interfaces'
+import create from './routes/create';
+import delete_ from './routes/delete';
+import get from './routes/get';
+import list from './routes/list';
+import patch from './routes/patch';
+import update from './routes/update';
 
-export { defineUserRoutes };
+export default prefixRoutes('/users', [
+  create,
+  delete_,
+  get,
+  list,
+  patch,
+  update,
+]);
