--- assets/express/src/api/v1/users/routes/delete-user.ts
@@ -1,28 +1,0 @@
-import { bindRouteOptionsWithSchema } from '../../../interfaces';
-import { emailUserDTO } from '../dto';
-import { response } from '@common';
-import { UserService } from '@modules';
-
-export default bindRouteOptionsWithSchema(
-  ({ userService }: { userService: UserService }) => ({
-    operationId: 'user-delete',
-    summary: 'Delete an User',
-    description: 'Soft delete an User item',
-    tags: ['User'],
-    method: 'delete',
-    path: '/:email',
-    request: {
-      params: emailUserDTO,
-    },
-    responses: {
-      204: response.NoContent(),
-      404: response.NotFound(),
-      422: response.UnprocessableEntity(),
-    },
-    handler: async function (req, res) {
-      await userService.delete(req.params.email);
-
-      res.status(204);
-    },
-  })
-);
